fix(propdefs): probe writability on every acquire, not on a schedule
sqlx's `test_before_acquire` already spends a round trip per acquire on a ping, so
`SHOW transaction_read_only` can replace that ping instead of being scheduled around
it. Probing every acquire then costs nothing extra, and it removes the heartbeat and
settle windows along with the blind spot between probes: a reader is now caught on the
next write rather than up to a heartbeat later.

Deliberately `before_acquire` only, not `after_connect`. A fresh connection that lands
on a reader serves one write and is caught next acquire, whereas an `after_connect`
error sends sqlx into reconnect-backoff until `acquire_timeout`, turning a
cluster-wide read-only state into `PoolTimedOut` instead of a 25006 the caller can
classify.

Also moves `error_class` and `is_read_only_error` next to the other SQLSTATE
classifiers in `lib.rs` — neither is writer-guard logic — and corrects the
`max_lifetime(None)` comment that claimed `test_before_acquire` covered connection
health. It covers a severed socket, not a server that stopped accepting writes.
